+++ qh3server/qh3server/http3_command_server.cpp
@@ -8,7 +8,7 @@
 #include "http3_command_server.hpp"
 #include "../../common/gxcrc32.h"
 
-http3_command_server::http3_command_server(const qstring& redis_ip, int redis_port, bridge_command_center* bridge_, qstring router_port_) : bridge(bridge_), router_port(router_port_) {
+http3_command_server::http3_command_server(const qstring& redis_ip, uint16_t redis_port, bridge_command_center* bridge_, qstring router_port_) : bridge(bridge_), router_port(router_port_) {
     hiredis = DEBUG_NEW qhiredis();
     hiredis->connect_redis(redis_ip, redis_port);
 }
